requirements.txt
coverage==4.5.3
Django==2.1.7
django-cors-headers==2.5.2
django-filter==2.1.0
django-polymorphic==2.0.3
django-rest-passwordreset==1.0.0
django-rest-polymorphic==0.1.8
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.1
drf-writable-nested==0.5.1
gunicorn==19.9.0
Pillow==6.0.0
psycopg2-binary==2.8.2
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
